Singer's attendance monitoring has been implemented.
An attendance is created for each singer with a voice when an event is created.
Added the ability to remove a singer from an event.
Singers can edit their attendance status.

Added separation of a large amount of data when creating buttons.
Some hard-coded text is moved to dictionaries.

Fixed saving the date when creating an event.
Fixed date and time editing.