Some cosmetics and bug fixes.
Button names moved into the buttons_dictionary.py
The repeat event option is only available for rehearsals.
Added editing of participants after creating an event.

Fixed a bug where the bot would send a DELETE message when the cancel button was clicked.
Fixed a bug with repeated sending of text when saving a new location while adding an event.
Fixed bug with empty attendance for new singers.
Fixed a bug with incorrect checking of the deleted participation of the singer.